requirements.txt
asgiref==3.8.1
carehare==1.0.2
channels==4.2.0
Django==5.1.7
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
drf-restwind==0.1.6
gunicorn==23.0.0
Markdown==3.7
msgpack==1.1.0
packaging==24.2
pamqp==3.3.0
pika==1.3.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
